Extract SessionMonitor as the single-writer observation path
Move the state-mutating observation core out of GovernancePipeline into a
dedicated SessionMonitor collaborator (governance/monitor.py): observe_event,
get_or_create_state, process_lifecycle, process_event, and the write helpers
_commit_mcp_writes_no_commit / _write_session_summary, plus the reservation
bookkeeping (_write_failures, _phase23_attempts, _phase23_session_keys and their
caps). SessionMonitor is now the only component that mutates + persists session
state; the Scorer owns the read/preview side.

The durability model is preserved verbatim: idempotency reservation, atomic
Phase-1 commit, event-time snapshot on retries, the Phase 2/3 circuit breaker /
dead-letter, deferred MCP writes, and retry-counter clearing only after a
successful finalization commit. The only edit to moved code is redirecting
observe_event's self.context_from_session_event -> self._context.from_session_event.

GovernancePipeline keeps thin forwards for the public API (observe_event,
process_event, process_lifecycle, get_or_create_state) so external callers
(EventPipeline stage, tests) are unaffected. Now-unused pipeline imports
(json, sqlite3, datetime/timezone, SessionStateSnapshot) removed.

Full suite green (1884 passed, 4 skipped); ruff clean.
